requirements.txt
langchain
openai
faiss-cpu
python-dotenv
pypdf
flask
docx2txt
